code clean up widget updates tests added
