+++ BobbysCounter/Entity/Counter.swift
