+++ BobbysCounter/Entity/Errors.swift
