+++ BobbysCounter/Entity/Selected.swift
